Initial CSE implementation, with cheated proofs
Initial CSE implementation, built on a value numbering analysis.
This is a restricted version of the optimisation, in the interest of
ensuring the proofs work in the context of word_to_word, modifying
only redundant moves and binops.

There are known issues with this version, primarily the use of a
locals relationship similar to locals_rel but without the upper bound on
vars. The lack of this restriction prevents reuse of several existing
locals_rel proofs.